+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Text in chat wraps at the end of the window
 - Initial Discord going-live integration
 - Add Discord frontend
+- Moved functionality from main to `ImGuiManager`.
+- Initial implementation of OBS WebSocket connection
 
 ### Changed
 - Added a default fallback font (FiraSans - random right now!)